Initial synth layout and inner working.
This synth will require a generator for everything but leaf nodes MidiIn
and Fixed. All other data must come from there.

Even LFOs need some tunning that must come ultimately from there.

Sound modules are ProcessBlocks, and are first added to the synth, and
then patched. This will allow to let the synth generate all the dynamic
memory required for all the buffers on the connections, and to schedule
the blocks as needed.

Blocks can have any number of inputs and outputs. Each connection will
be an audiobuffer allocated at the beginning and reused at each loop.

Ultimately the synths are generated from a simple synth definition, and
used with no UI, all MIDI controlled. Ideally from a raspberry Pi with
just a MIDI keyboard connected.