Audiobuffers can be used to manipulate samples
They are separated in two parts, the input and the output. Input are RO,
output a RW.

The idea is to have the work order for each synth step in a struct with
the reference to the module, the input audio buffers, and the output
audiobuffers. This order is then just executed bindly, no allocations
required, and use the final output buffer as output data.

For this to work the audio buffers are shared always in one output
(from the generator, or prefilled), and one input, using refcells.

Also experiment to use YAML instead of custom format, as may be easier
to just parse.