Added in/out support by vec on synth
Create two audiobuffer vectors (input/output), each with diferent
behavours, but sharing audiobuffers, so that the output of one
module is the input of another. This way when we order proeprly the
blocks, we can do the work just calling each module in the apporpiate
order, and use the output buffer.