Added monophonic note return to prev
So that if a note is released and one was pressed from before goes
there. There is a FIFO buffer of 8 notes.

For reference MiniMoog was higher note rules, so you could press first
high pitch, then low pich and high pith continues sounding, release high
pitch and low starts to sound.

This makes it sound in order and at release any of the previous in
order.